refactor: split ingestion replay dtos
